requirements.txt
pandas
openai
python-dotenv
matplotlib
pymupdf
langchain
arxiv
tiktoken
tqdm